SCRUM-30 add UT for fetch board
