+++ src/lib/game.tsx
@@ -117,27 +117,3 @@ export async function start_game({ game_id, player_id }: { game_id: number, play
     };
   }
 }
-
-export async function fetch_board({ game_id }: { game_id: number }) {
-  try {
-    const response = await fetch(`http://localhost:8000/game/board`, {
-      method: 'GET',
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ "game_id": game_id }),
-    });
-
-    if (!response.ok) {
-      throw new Error(`Server responded with status ${response.status}`);
-    }
-
-    const result = await response.json();
-
-    return result;
-
-  } catch (error) {
-    console.error('Failed to fetch board:', error);
-    return { status: 'ERROR', message: 'An error occurred while fetching the board' };
-  }
-}
